python/requirements.txt
fastapi
uvicorn
geopandas
numpy
shapely
pydantic
matplotlib
